# Windows debug info lookup improvements