These are some changes to reduce the amount of kernel32 calls, file reads,  and allocations required to render stack traces on windows.

The previous method would do the following for each address in the trace:

1. Two kernel32 calls to get the module list
2. For each module (until the one with the address in it is found)
  1. kernel32 call to get the module info
  2. If it's the correct module:
    - kernel32 call to get it's filename
    - Read the file from disk and parse Coff

Instead, now when the debug info is created, we make a call to `kernel32.CreateToolhelp32Snapshot` to snapshot the module list. The set of module sizes/base addrs is then copied. 

Then, the address -> module lookup changes to:
1. Iterate the cache to find the base addr of the module
2. Using the base address, use the Coff parser to read the module from memory directly (since it's already mapped into our address space)

This saves some disk I/O reading the dll/exe from disk, as well as reducing the number of module list lookups from N to 1. It also fixes a TOCTOU issue that existed in `lookupModuleWin32`.

Other changes:
- `unexpectedStatus`/`unexpectedError` now start the stack trace at the caller instead of inside the debug machinery
- Free DWARF debug info that's allocated in `readCoffDebugInfo`


### Benchmark

Generate 256 stack traces per run, traversing 3 DLLs.

Source here: https://gist.github.com/kcbanner/6b2473a9d5fdd625f021d3429c3b2239

```
"Master version"
Benchmark 1: zig-out.master\bin\zig_panic.exe
  Time (mean ± σ):     996.0 ms ±  14.6 ms    [User: 288.4 ms, System: 703.8 ms]
  Range (min … max):   977.2 ms … 1031.6 ms    10 runs

"New version"
Benchmark 1: zig-out.new\bin\zig_panic.exe
  Time (mean ± σ):     636.9 ms ±   8.2 ms    [User: 220.3 ms, System: 420.0 ms]
  Range (min … max):   627.7 ms … 657.9 ms    10 runs
```

### Note about syscalls

Note: I've swapped some kernel32 calls for other kernel32 calls here. I know the goal is to rely on only ntdll, however there is an issue with this in this case. It is possible to replace `lookupModuleWin32` with something like the following snippet:

```zig
    fn lookupModuleWin32(self: *DebugInfo, address: usize) !*ModuleDebugInfo {
        const teb = std.os.windows.teb();
        const peb = teb.ProcessEnvironmentBlock;

        if (!builtin.single_threaded) {
            // TODO: Verify the offset of this field on x86 and x64
            windows.kernel32.EnterCriticalSection(peb.LoaderLock);
        }

        var iter = peb.Ldr.InMemoryOrderModuleList.Flink;
        while (iter != &peb.Ldr.InMemoryOrderModuleList) : (iter = iter.Flink) {
            const table_entry = @fieldParentPtr(windows.LDR_DATA_TABLE_ENTRY, "InMemoryOrderLinks", iter);
            const base = @ptrToInt(table_entry.DllBase);

            if (address >= base and address < base + table_entry.SizeOfImage) {
                if (self.address_map.get(base)) |obj_di| {
                    return obj_di;
                }

                const mapped_module = @ptrCast([*]const u8, table_entry.DllBase)[0..table_entry.SizeOfImage];
                const obj_di = try self.allocator.create(ModuleDebugInfo);
                errdefer self.allocator.destroy(obj_di);

                obj_di.* = try readCoffDebugInfo(self.allocator, mapped_module);
                obj_di.base_address = base;

                try self.address_map.putNoClobber(base, obj_di);
                return obj_di;
            }
        }

        if (!builtin.single_threaded) {
            windows.kernel32.LeaveCriticalSection(peb.LoaderLock);
        }

        return error.MissingDebugInfo;
    }
```

However, the problem with this is that it relies on `peb.LoaderLock`, and according to https://www.geoffchappell.com/studies/windows/km/ntoskrnl/inc/api/pebteb/peb/index.htm, the offset of this field changes based on the windows version. There is a comment about this in the `windows.zig` version of this struct. So, if we only have one definition of the PEB struct, then this would break on versions of windows that don't match up with the layout we've picked.

Other options would include re-implementing what `CreateToolhelp32Snapshot` does manually  Based on ReactOS source (https://doxygen.reactos.org/d2/d21/dll_2win32_2kernel32_2client_2toolhelp_8c.html#a3a41581a0fbf9490c17bad7503c9b732) it calls more undocumented ntdll functions. I'm open to implementing this, if we are ok relying on these undocumented functions.