@andrewrk 2023-01-10
Thanks for this!

> I'm open to implementing this, if we are ok relying on these undocumented functions.

At this point we are fully in the business of relying on undocumented NtDll functions. I have a plan to address this if it ever becomes a problem in practice.